Be explicit about what URL to configure for OAuth
We're telling people to create a GitHub OAuth application that points to their app, but
we've not said anything about what URL to put there.

I'm assuming that this URL is used for the redirect after logging in and authorizing the application,
in which case the localhost:3000 is fine as a default.

If we need to process web hooks, then we will have to instruct people to set up ngrok or localtunnel.