List pull requests on user page by created date
The /users/:username page was showing a list of dates, many of which
were not Fridays.

Upon inspection it turned out that all of these were created on a
Friday, but they were being listed by the date that they were closed or
merged.

Since we're specifically calling out Fridays with the heading

> N pull requests opened on Fridays in the last year:

it feels confusing to have non-Fridays listed.

It also seems a bit odd to show them by the close/merge date rather
than the open date.