rescue and handle Octokit::Unauthorized
this error can occur when the current user's oauth token is not valid.
as far as I can tell this is the only code that uses the token.

when the token is invalid, sign the current user out and send them back to
the homepage with a flash error message prompting them to sign in