Refactor directory structure, update _config.yml to match opensource.guide
This change updates the directory structure of the jekyll site, moving
the language-specific directories under `_articles` and updating the
front matter / `_config.yml` to match.

The goal was to clean up the root of the repository and bring the
config in line with that used by the opensource.guide site.